Update link in event.md to point to master version
Before this change, the link to the inline docs pointed to an older version of ReactEventRe.rei. This version of the file used the older Reason syntax. This change updates the link to point to the version of this file in master. This means that this issue will not happen as the link will always point to the latest version of the file that is in master.