Update example to working component.
The current documentation uses the deprecated `createClass` method, which throws an error.  
Rather than adding another dependency, it seems easier to use a stateless react component.  

Additionally, you can't return null from a component, so if you were to test `show=false`, you will get an error.  So to minimize potential confusion, I updated to return with true or false.

credit to: https://reasonml.chat/t/typeerror-react-createclass-is-not-a-function/388/6